+++ src/main/kotlin/domain/Car.kt
@@ -1,4 +1,4 @@
-package model
+package domain
 
 data class Car(val name: String) {
     var position: Int = CAR_INITIAL_POSITION
