+++ src/main/kotlin/domain/Game.kt
@@ -1,4 +1,4 @@
-package model
+package domain
 
 import java.lang.IllegalArgumentException
