+++ src/test/kotlin/CarTest.kt
@@ -1,4 +1,4 @@
-import model.Car
+import domain.Car
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
